+++ BacktestEngine/src/execution/fee_model.h
@@ -7,13 +7,13 @@ class IFeeModel
 {
 public:
     virtual ~IFeeModel() = default;
-    virtual double compute_commission(order_side side, int quantity, double price) const = 0;
+    virtual double compute_commission(order_side side, double quantity, double price) const = 0;
 };
 
 class ZeroFeeModel : public IFeeModel
 {
 public:
-    double compute_commission(order_side /*side*/, int /*quantity*/, double /*price*/) const override
+    double compute_commission(order_side /*side*/, double /*quantity*/, double /*price*/) const override
     {
         return 0.0;
     }
@@ -24,7 +24,7 @@ class FixedFeeModel : public IFeeModel
 public:
     explicit FixedFeeModel(double fee_per_trade) : fee_per_trade_(fee_per_trade) {}
 
-    double compute_commission(order_side /*side*/, int /*quantity*/, double /*price*/) const override
+    double compute_commission(order_side /*side*/, double /*quantity*/, double /*price*/) const override
     {
         return fee_per_trade_;
     }
@@ -39,7 +39,7 @@ public:
     TieredFeeModel(double maker_rate, double taker_rate)
         : maker_rate_(maker_rate), taker_rate_(taker_rate) {}
 
-    double compute_commission(order_side side, int quantity, double price) const override
+    double compute_commission(order_side side, double quantity, double price) const override
     {
         double notional = std::abs(quantity * price);
         // Buys are takers (crossing the spread), sells are makers (providing liquidity)
